chore: rename MaxComputeSchemaHelper.java to MaxComputeSchemaBuilder
